+++ apps/studio/src/schemas/redirect/redirect.ts
@@ -2,8 +2,13 @@ import { REFERENCE_LINK_REGEX } from "@opengovsg/isomer-components"
 import { z } from "zod"
 import { RESERVED_SOURCE_PREFIXES } from "~/constants/redirect"
 
-import { generateBigIntSchema } from "./common"
-import { offsetPaginationSchema } from "./pagination"
+import { generateBigIntSchema } from "../common"
+import { offsetPaginationSchema } from "../pagination"
+import {
+  normalizeRedirectPath,
+  normalizeRedirectSource,
+  trimSlashes,
+} from "./utils"
 
 // A redirect is published as an S3 object whose key is
 // `${siteName}/${buildNumber}/latest/${source}/index.html` (see
@@ -66,22 +71,6 @@ export const isValidExternalDestination = (value: string) => {
     return false
   }
 }
-
-// Strips slashes from both ends of a path so "/foo/", "foo" and "foo//"
-// all normalise to the same inner segments before validation.
-const trimSlashes = (value: string) => value.replace(/^\/+|\/+$/g, "")
-
-// Normalises a path to a single leading slash, no trailing slash, collapsed
-// runs ("/foo/", "foo", "foo//" -> "/foo"). Exported so the server can compare
-// a destination path against stored sources, persisted in this form.
-export const normalizeRedirectPath = (value: string) =>
-  `/${trimSlashes(value).replace(/\/{2,}/g, "/")}`
-
-// Sources are additionally lowercased — page permalinks are lowercase-only, so
-// a source must lowercase to compare against (and not shadow) a real page.
-// Exported so the server's source/loop guards compare in the same form.
-export const normalizeRedirectSource = (value: string) =>
-  normalizeRedirectPath(value).toLowerCase()
 
 // True when the (normalised) source falls under a reserved prefix — the prefix
 // itself or anything nested beneath it.
