+++ apps/studio/src/server/modules/redirect/redirect.service.ts
@@ -30,10 +30,12 @@ import {
 import { parseRedirectCsv } from "~/lib/redirectCsv"
 import {
   isValidExternalDestination,
-  normalizeRedirectPath,
-  normalizeRedirectSource,
   redirectRowSchema,
 } from "~/schemas/redirect"
+import {
+  normalizeRedirectPath,
+  normalizeRedirectSource,
+} from "~/schemas/redirect/utils"
 import { getReferenceLink } from "~/utils/link"
 import { ResourceType } from "~prisma/generated/generatedEnums"
