+++ apps/studio/src/server/modules/resource/resource.service.ts
@@ -15,7 +15,7 @@ import { INDEX_PAGE_PERMALINK } from "~/constants/sitemap"
 import {
   normalizeRedirectPath,
   normalizeRedirectSource,
-} from "~/schemas/redirect"
+} from "~/schemas/redirect/utils"
 import {
   getSitemapTree,
   injectTagMappings,
